Integrate the C compiler to the simulator
