+++ cucu/cucu.c
