+++ cucu/gen.c
