--- from_yyk/readme.txt
@@ -1,5 +1,0 @@
-cucu.py, cucu-dummy.exe, Simserver.pys三个文件应放置在与index.html同一目录下
-启动server: Python Simserver.py 
-http://localhost:8080/index.html 为目前的页面url
-在textarea内编辑c 代码，点击complie返回v8代码，点击assemble后可以run,简单写了两个testcase，是test1.txt和test2.txt。
-
