--- from_yyk/test1.txt
@@ -1,6 +1,0 @@
-int main()
-{
-int i = 3;
-int j = 5;
-i = i + j;
-}
