--- from_yyk/test2.txt
@@ -1,8 +1,0 @@
-int main()
-{
-    int i = 5;
-    while(i != 10)
-    {
-        i = i + 1;
-    }
-}
