--- from_yyk/test3.txt
@@ -1,9 +1,0 @@
-int main()
-{
-    int i = 3;
-    int j = 5;
-    if(i != 5)
-    {
-        j = 6;
-    }
-}
