+++ index.html
@@ -28,14 +28,27 @@
             <div class="col-lg-7 col-md-6">
                 <div class="panel panel-default">
                     <div class="panel-heading">
-                      <h4 class="panel-title">Code
-			<small>(<a href="http://enight.me/files/csao/v8spec.pdf" target="_blank" style="color: #337AB7">Architecture Specification</a>&nbsp;&nbsp;<a href="./instruction-set.html" target="_blank" style="color: #337AB7">Instruction Set</a>)
-		      </small></h4>
+                      <h4 class="panel-title">Code <small>(<a href="./instruction-set.html" target="_blank" style="color: #337AB7">Instruction Set</a>) (<a href="http://enight.me/files/csao/v8spec.pdf" target="_blank" style="color: #337AB7">Specification</a>)</small></h4>
                     </div>
-                    <button type="button" class="btn btn-default" ng-click="assemble()">Assembly</button>
-                    <button type="button" class="btn btn-default" ng-click="upload()">Machine Code</button>
+                    <button type="button" class="btn btn-default" ng-click="compile()">Compile</button>
+                    <script>
+	function postcode()
+    {
+        var ta = document.getElementById("sourceCode");
+        $.post('/',{"1":ta.value},function(response){
+        ta.value = response;
+	});
+    }
+
+</script>
+                    <button type="button" class="btn btn-default" ng-click="assemble()">Assemble</button>
+                    <button type="button" class="btn btn-default" ng-click="upload()">Upload</button>
 		    Examples:
                     <select ng-model="example" ng-options="item.id as item.desc for item in examples" ng-change="showExample(example)" ng-init="initExamples()"></select>
+		    <!-- <select id="examples" ng-change="showExample(x)" ng-model="x"> -->
+		    <!--   <option value="haha">haha</option> -->
+		    <!--   <option value="hoho">hoho</option> -->
+		    <!-- </select> -->
                     <form role="form">
                       <textarea id="sourceCode"
                                 class="form-control source-code"
@@ -46,15 +59,15 @@
                                 ng-model="code"></textarea>
                     </form>
                 </div>
+            </div>
+            <div class="clearfix visible-xs visible-sm"></div>
+            <div class="col-lg-5 col-md-6">
 		<div class="panel panel-default">
                     <div class="panel-heading">
                         <h4 class="panel-title">Printer</h4>
                     </div>
 		    <div class="panel-body source-code">{{ printer.data }}</div>
 		</div>
-            </div>
-            <div class="clearfix visible-xs visible-sm"></div>
-            <div class="col-lg-5 col-md-6">
                 <div class="panel panel-default">
                     <div class="panel-heading">
                         <h4 class="panel-title">CPU & Memory</h4>
@@ -171,5 +184,7 @@
     </div>
     <script src="node_modules/angular/angular.min.js"></script>
     <script src="assets/asmsimulator.min.js"></script>
+    <script src="http://code.jquery.com/jquery-1.11.3.min.js"></script>
+
 </body>
 </html>
